fix: redirect import of definePageConfig
