+++ .changeset/rotten-adults-try.md
@@ -1,0 +1,6 @@
+---
+'@ice/shared-config': patch
+'@ice/app': patch
+---
+
+fix: redirect import of definePageConfig
