+++ packages/ice/src/plugins/task.ts
@@ -1,19 +1,12 @@
 import * as path from 'path';
 import { createRequire } from 'module';
 import type { Config } from '@ice/shared-config/types';
-import { CACHE_DIR, RUNTIME_TMP_DIR, WEB } from '../../constant.js';
+import { CACHE_DIR, RUNTIME_TMP_DIR } from '../constant.js';
 
 const require = createRequire(import.meta.url);
-const getWebTask = ({ rootDir, command, userConfig }): Config => {
+const getDefaultTaskConfig = ({ rootDir, command }): Config => {
   // basic task config of web task
   const defaultLogging = command === 'start' ? 'summary' : 'summary assets';
-  const removeExportExprs = ['serverDataLoader', 'staticDataLoader'];
-
-  // Remove dataLoader exports only when build in production
-  // and configure to generate data-loader.js.
-  if (command === 'build' && userConfig.dataLoader) {
-    removeExportExprs.push('dataLoader');
-  }
 
   const envReplacement = path.join(rootDir, RUNTIME_TMP_DIR, 'env.ts');
   return {
@@ -22,6 +15,7 @@ const getWebTask = ({ rootDir, command, userConfig }): Config => {
     cacheDir: path.join(rootDir, CACHE_DIR),
     alias: {
       ice: path.join(rootDir, RUNTIME_TMP_DIR, 'index.ts'),
+      'ice/types': path.join(rootDir, RUNTIME_TMP_DIR, 'types.ts'),
       '@': path.join(rootDir, 'src'),
       // set alias for webpack/hot while webpack has been prepacked
       'webpack/hot': '@ice/bundles/compiled/webpack/hot',
@@ -32,17 +26,13 @@ const getWebTask = ({ rootDir, command, userConfig }): Config => {
       'universal-env': envReplacement,
       '@uni/env': envReplacement,
     },
-    swcOptions: {
-      removeExportExprs,
-    },
+
     assetsManifest: true,
     fastRefresh: command === 'start',
     logging: process.env.WEBPACK_LOGGING || defaultLogging,
     minify: command === 'build',
     useDevServer: true,
-    useDataLoader: true,
-    target: WEB,
   };
 };
 
-export default getWebTask;
+export default getDefaultTaskConfig;
