+++ packages/ice/src/service/config.ts
@@ -1,5 +1,6 @@
 import * as path from 'path';
 import fs from 'fs-extra';
+import type { Config as SharedConfig } from '@ice/shared-config/types';
 import type { ServerCompiler } from '../types/plugin.js';
 import { getCache, setCache } from '../utils/persistentCache.js';
 import { getFileHash } from '../utils/hash.js';
@@ -9,6 +10,7 @@ import { RUNTIME_TMP_DIR, CACHE_DIR } from '../constant.js';
 import { createLogger } from '../utils/logger.js';
 import externalBuiltinPlugin from '../esbuild/externalNodeBuiltin.js';
 
+
 type GetOutfile = (entry: string, exportNames: string[]) => string;
 
 interface CompileConfig {
@@ -17,6 +19,7 @@ interface CompileConfig {
   transformInclude?: (id: string) => boolean;
   needRecompile?: (entry: string, options: string[]) => Promise<boolean | string>;
   getOutfile?: GetOutfile;
+  redirectImports?: SharedConfig['redirectImports'];
 }
 
 class Config {
@@ -39,7 +42,7 @@ class Config {
 
   public setCompiler(serverCompiler: ServerCompiler): void {
     this.compiler = async (keepExports) => {
-      const { entry, transformInclude } = this.compileConfig;
+      const { entry, transformInclude, redirectImports } = this.compileConfig;
       const outfile = this.getOutfile(entry, keepExports);
       this.status = 'PENDING';
       const { error } = await serverCompiler({
@@ -61,6 +64,7 @@ class Config {
             include: transformInclude,
           },
         },
+        redirectImports,
       });
       if (!error) {
         this.status = 'RESOLVED';
@@ -211,6 +215,10 @@ export const getRouteExportConfig = (rootDir: string) => {
         return false;
       }
     },
+    redirectImports: [{
+      specifier: ['definePageConfig'],
+      source: 'ice/types',
+    }],
   });
 
   const dataloaderConfig = new Config({
