+++ packages/ice/templates/core/entry.server.ts.ejs
@@ -1,7 +1,7 @@
 import './env.server';
 import * as runtime from '@ice/runtime/server';
 <% if(hydrate) {-%>
-import { commons, statics } from './runtimeModules';
+import { commons, statics } from './runtime-modules';
 <% }-%>
 import * as app from '@/app';
 import * as Document from '@/document';
