+++ packages/ice/templates/core/runtime-modules.ts.ejs
